+++ contract/population_types.go
@@ -1,4 +1,4 @@
-package models
+package contract
 
 type PopulationTypes struct {
 	Items []PopulationType `json:"items"`
